requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
apscheduler
beautifulsoup4
supabase
